+++ src/components/auth/Login.js
@@ -1,61 +1,49 @@
 import React, { useRef } from "react"
 import { Link, useHistory } from "react-router-dom"
-import "./Login.css"
+// import "./Auth.css"
 
-export const Login = (props) => {
+export const Login = () => {
   const email = useRef()
   const password = useRef()
-  const existDialog = useRef()
-  const passwordDialog = useRef()
+  const invalidDialog = useRef()
   const history = useHistory()
-
-  const existingUserCheck = () => {
-    return fetch(`http://localhost:8088/users?email=${email.current.value}`)
-      .then((_) => _.json())
-      .then((user) => ("id" in user ? user : false))
-  }
-
-  const getall = () => {
-    return fetch(`http://localhost:8088/users`).then((res) =>
-      res.json().then((users) => console.log(users))
-    )
-  }
 
   const handleLogin = (e) => {
     e.preventDefault()
-    getall()
-    existingUserCheck().then((exists) => {
-      console.log(exists)
-      console.log(email)
-      console.log(password)
-      if (exists && exists.password === password.current.value) {
-        localStorage.setItem("rare_user_id", exists.id)
-        history.push("/")
-      } else if (exists && exists.password !== password.current.value) {
-        passwordDialog.current.showModal()
-      } else if (!exists) {
-        existDialog.current.showModal()
-      }
+
+    return fetch("http://127.0.0.1:8088/login", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+        "Accept": "application/json",
+      },
+      body: JSON.stringify({
+        email: email.current.value,
+        password: password.current.value,
+      }),
     })
+      .then((res) => res.json())
+      .then((res) => {
+        if ("valid" in res && res.valid) {
+          localStorage.setItem("rare_user_id", res.token)
+          history.push("/")
+        } else {
+          invalidDialog.current.showModal()
+        }
+      })
   }
 
   return (
     <main className="container--login">
-      <dialog className="dialog dialog--auth" ref={existDialog}>
-        <div>User does not exist</div>
-        <button className="button--close" onClick={(e) => existDialog.current.close()}>
-          Close
-        </button>
-      </dialog>
-      <dialog className="dialog dialog--password" ref={passwordDialog}>
-        <div>Password does not match</div>
-        <button className="button--close" onClick={(e) => passwordDialog.current.close()}>
+      <dialog className="dialog dialog--auth" ref={invalidDialog}>
+        <div>Email or password was not valid.</div>
+        <button className="button--close" onClick={(e) => invalidDialog.current.close()}>
           Close
         </button>
       </dialog>
       <section>
         <form className="form--login" onSubmit={handleLogin}>
-          <h1>Nashville Kennels</h1>
+          <h1>Level Up</h1>
           <h2>Please sign in</h2>
           <fieldset>
             <label htmlFor="inputEmail"> Email address </label>
@@ -64,6 +52,7 @@ export const Login = (props) => {
               type="email"
               id="email"
               className="form-control"
+              defaultValue="me@me.com"
               placeholder="Email address"
               required
               autoFocus
@@ -76,12 +65,19 @@ export const Login = (props) => {
               type="password"
               id="password"
               className="form-control"
+              defaultValue="me"
               placeholder="Password"
               required
             />
           </fieldset>
-          <fieldset>
-            <button type="submit">Sign in</button>
+          <fieldset
+            style={{
+              textAlign: "center",
+            }}
+          >
+            <button className="btn btn-1 btn-sep icon-send" type="submit">
+              Sign In
+            </button>
           </fieldset>
         </form>
       </section>
